Module pallet_order_book::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

add_trading_pair

Adds a valid trading pair.

create_order

Create an order with the default min fulfillment amount.

fill_order_full

Fill an existing order, fulfilling the entire order.

fill_order_partial

Fill an existing order, based on the provided partial buy amount.

rm_trading_pair

Removes a valid trading pair

update_min_order

Sets the minimum order amount for a given trading pair. If the trading pair is not yet added this errors out.

user_cancel_order

Cancel an existing order that had been created by calling account.

user_update_order

Update an existing order