Trait pallet_order_book::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type AssetRegistry: Inspect<AssetId = Self::AssetCurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata>;
    type AssetCurrencyId: AssetId + Parameter + Debug + Default + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen;
    type OrderIdNonce: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + EnsureAdd + MaybeSerializeDeserialize + TypeInfo + MaxEncodedLen;
    type Balance: Member + Parameter + FixedPointOperand + BaseArithmetic + EnsureMul + EnsureDiv + TypeInfo + MaxEncodedLen;
    type TradeableAsset: AssetInspect<Self::AccountId, Balance = Self::Balance, AssetId = Self::AssetCurrencyId> + InspectHold<Self::AccountId, Reason = ()> + MutateHold<Self::AccountId> + Mutate<Self::AccountId>;
    type SellRatio: Parameter + Member + FixedPointNumber + EnsureMul + EnsureDiv + MaybeSerializeDeserialize + TypeInfo + MaxEncodedLen;
    type OrderPairVecSize: Get<u32>;
    type MinFulfillmentAmountNative: Get<Self::Balance>;
    type DecimalConverter: ConversionToAssetBalance<Self::Balance, Self::AssetCurrencyId, Self::Balance, Error = DispatchError>;
    type FulfilledOrderHook: StatusNotificationHook<Id = Self::OrderIdNonce, Status = Swap<Self::Balance, Self::AssetCurrencyId>, Error = DispatchError>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Weights: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type AssetRegistry: Inspect<AssetId = Self::AssetCurrencyId, Balance = Self::Balance, CustomMetadata = CustomMetadata>

Asset registry for foreign currencies we can take orders for.

type AssetCurrencyId: AssetId + Parameter + Debug + Default + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen

CurrencyId of Assets that an order can be made for

type OrderIdNonce: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + EnsureAdd + MaybeSerializeDeserialize + TypeInfo + MaxEncodedLen

Type used for OrderId. OrderIdNonce ensures each OrderId is unique. OrderIdNonce incremented with each new order.

type Balance: Member + Parameter + FixedPointOperand + BaseArithmetic + EnsureMul + EnsureDiv + TypeInfo + MaxEncodedLen

Balance type

type TradeableAsset: AssetInspect<Self::AccountId, Balance = Self::Balance, AssetId = Self::AssetCurrencyId> + InspectHold<Self::AccountId, Reason = ()> + MutateHold<Self::AccountId> + Mutate<Self::AccountId>

Type for currency orders can be made for

type SellRatio: Parameter + Member + FixedPointNumber + EnsureMul + EnsureDiv + MaybeSerializeDeserialize + TypeInfo + MaxEncodedLen

Type for price ratio for cost of incoming currency relative to outgoing

type OrderPairVecSize: Get<u32>

Size of order id bounded vec in storage

type MinFulfillmentAmountNative: Get<Self::Balance>

The default minimum fulfillment amount for orders.

NOTE: The amount is expected to be denominated in native currency. When applying to a swap order, it will be re-denominated into the target currency.

type DecimalConverter: ConversionToAssetBalance<Self::Balance, Self::AssetCurrencyId, Self::Balance, Error = DispatchError>

Type which provides a decimal conversion from native to another currency.

NOTE: Required for MinFulfillmentAmountNative.

type FulfilledOrderHook: StatusNotificationHook<Id = Self::OrderIdNonce, Status = Swap<Self::Balance, Self::AssetCurrencyId>, Error = DispatchError>

The hook which acts upon a (partially) fulfilled order

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

The admin origin of this pallet

type Weights: WeightInfo

Type for pallet weights