Type Alias pallet_order_book::pallet::OrderOf

pub type OrderOf<T> = Order<<T as Config>::OrderIdNonce, <T as Config>::AccountId, <T as Config>::AssetCurrencyId, <T as Config>::Balance, <T as Config>::SellRatio>;

Order of pallet config type

Aliased Type

struct OrderOf<T> {
    pub order_id: <T as Config>::OrderIdNonce,
    pub placing_account: <T as Config>::AccountId,
    pub asset_in_id: <T as Config>::AssetCurrencyId,
    pub asset_out_id: <T as Config>::AssetCurrencyId,
    pub buy_amount: <T as Config>::Balance,
    pub initial_buy_amount: <T as Config>::Balance,
    pub max_sell_rate: <T as Config>::SellRatio,
    pub min_fulfillment_amount: <T as Config>::Balance,
    pub max_sell_amount: <T as Config>::Balance,
}

Fields
order_id: <T as Config>::OrderIdNonceplacing_account: <T as Config>::AccountIdasset_in_id: <T as Config>::AssetCurrencyIdasset_out_id: <T as Config>::AssetCurrencyIdbuy_amount: <T as Config>::Balance

How many tokens of asset in to buy

initial_buy_amount: <T as Config>::Balance

Original buy amount, used for tracking amount fulfilled

max_sell_rate: <T as Config>::SellRatio

Maximum relative price of the asset in being purchased relative to asset out ie: Rate::checked_from_rational(3u32, 2u32) would mean that 1 asset in would correspond with 1.5 asset out.

min_fulfillment_amount: <T as Config>::Balance

Minimum amount of an order that can be fulfilled for partial fulfillment

max_sell_amount: <T as Config>::Balance

Maximum amount of outgoing currency that can be sold